Bug 40418: Update the item type translation process to avoid Greybox modal
This patch updates the item type translation process so that it takes
place on a separate standalone page instead of in a Greybox modal.

Although currently only item types can be translated in this way, the
original plan was to expand the range of data that could be translated.
For this reason I used [% SWITCH %] in localization.tt for determining
whether to use the term "item type." This accommodates, I hope,
expansion of the feature.

To test you must have more than one installed translation.

- Apply the patch and go to Administration -> Item types.
- Edit an item type.
- On the "Description" line, click "Translate into other languages."
- You should be redirected to a new page, "Add translations for item
  type 'XXX.'
- Confirm that the page title, breadcrumbs, and heading are correct.
- Test the process of adding and deleting a translation.
- Test the process of performing an inline edit of an existing
  translation by clicking the table cell.
- Test that the "Return to item type 'XXX'" button returns you to the
  item type you were editing.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>