Bug 34978: Simplify tests logic
Joining the plugin classes strings and then performing regex checks on
it might be a bit fragile. I had hard time understanding the ands the
conditions for truth.

I rewrote it in a more intuitive way just using `none` and the tests are
easy to understand and pass.

The failure is probably because new plugins were added that make the
resulting string not match the expected structure.

The new new tests design is more readable anyway and not subject to this
errors in future additions.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>