Bug 40606: Remove italics from shelving location in the staff interface
This patch removes the italics formatting from the shelving location
now that it is in its own column. The formatting was there to
differentiate the shelving location from the branch name when they were
in the same column. This is not necessary anymore.

To test :
1. Go to a record with items which have a shelving location.
   --> The shelving location is in italics
2. Apply patch
3. `yarn build`
4. Refresh the page (Ctrl + F5)
   --> The shelving location is displayed normally

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
(cherry picked from commit 8451fafb0152c87609fc1350a087af421086e45a)
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>